requirements.txt
asgiref==3.9.1
certifi==2025.7.14
cftime==1.6.4.post1
contourpy==1.3.3
cycler==0.12.1
decouple==0.0.7
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
fonttools==4.59.0
kiwisolver==1.4.8
matplotlib==3.10.5
mysqlclient==2.2.7
netCDF4==1.7.2
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.16.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
xarray==2025.7.1
